Added Lambda Layer to hold ImageHash Python library
The ImageHash image processing library is too large to push up
to AWS with every change of our Lambda's implementations.  I've
separated it out into its own Lambda Layer and created the
necessary hookups in the stack definition to reference it.

I've also included a Dockerfile that can be used to generate the
Layer in case someone wants to do this later or wants to see
an example of how to do it.