Change detection Lambdas to publish on update_spam_score SNS topic (#5)
To solve an issue with the demo script, we decided to have the detection Lambdas publish their updates to the spam score to an SNS topic that the update_spam_score Lambda reads, instead of calling update_spam_score directly.

This solves the problem of the bad detect_spammy_words Lambda's error rate from increasing when the update_spam_score it invokes fails. By going through a queue, the detect_spammy_words Lambda will not know it incorrectly formated the request.

With this change, I also took the time to cleanup and refactor the code to eliminate a lot of duplicate code.